Fix some issues found by unit tests
